+++ src/main/java/com/github/seqware/BaseOxoGWrapperWorkflow.java
@@ -6,6 +6,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.function.BiConsumer;
 import java.util.function.Function;
+import java.util.stream.Collectors;
 
 import com.github.seqware.OxoGWrapperWorkflow.DownloadMethod;
 import com.github.seqware.OxoGWrapperWorkflow.Pipeline;
@@ -56,10 +57,12 @@ public abstract class BaseOxoGWrapperWorkflow extends AbstractWorkflowDataModel 
 	protected String broadGnosID;
 	protected String dkfzemblGnosID;
 	protected String museGnosID;
+	//NOTE: smufin does not have GNOS IDs.
 	
 	//This could be used implement a sort of local-file mode. For now, it's just used to speed up testing.
 	protected boolean skipDownload = false;
-	protected boolean skipUpload = false;
+	protected boolean skipBamUpload = false;
+	protected boolean skipVcfUpload = false;
 	
 	//Path to reference file usd for normalization, *relative* to /refdata/
 	protected String refFile = "public/Homo_sapiens_assembly19.fasta";
@@ -86,7 +89,10 @@ public abstract class BaseOxoGWrapperWorkflow extends AbstractWorkflowDataModel 
 	
 	protected final String workflowURL = "https://github.com/ICGC-TCGA-PanCancer/OxoGWrapperWorkflow/";
 	
-	protected String downloadMethod = "icgcStorageClient";
+	protected String vcfDownloadMethod = "icgcStorageClient";
+	protected String bamDownloadMethod;
+	
+	protected Map<Pipeline, DownloadMethod> pipelineDownloadMethods = new HashMap<OxoGWrapperWorkflow.Pipeline, OxoGWrapperWorkflow.DownloadMethod>(4);
 	
 	protected String sangerGNOSRepoURL;
 	protected String broadGNOSRepoURL;
@@ -115,6 +121,8 @@ public abstract class BaseOxoGWrapperWorkflow extends AbstractWorkflowDataModel 
 	protected boolean allowMissingFiles = false;
 	
 	List<VcfInfo> vcfs;
+	
+	String fileSystemSourcePath ;
 
 	/**
 	 * Get a property name that is mandatory
@@ -178,8 +186,66 @@ public abstract class BaseOxoGWrapperWorkflow extends AbstractWorkflowDataModel 
 		} ;
 
 		try {
-			if (hasPropertyAndNotNull("downloadMethod")) {
-				this.downloadMethod = getProperty("downloadMethod");
+			if (hasPropertyAndNotNull("skipDownload")) {
+				this.skipDownload = Boolean.valueOf(getProperty("skipDownload"));
+			}
+			
+			if (hasPropertyAndNotNull("skipBamUpload")) {
+				this.skipBamUpload = Boolean.valueOf(getProperty("skipBamUpload"));
+			}
+			
+			if (hasPropertyAndNotNull("skipVcfUpload")) {
+				this.skipVcfUpload = Boolean.valueOf(getProperty("skipVcfUpload"));
+			}
+			
+			if (hasPropertyAndNotNull("skipOxoG")) {
+				this.skipOxoG = Boolean.valueOf(getProperty("skipOxoG"));
+			}
+			
+			if (hasPropertyAndNotNull("skipVariantBam")) {
+				this.skipVariantBam = Boolean.valueOf(getProperty("skipVariantBam"));
+			}
+			
+			if (hasPropertyAndNotNull("skipAnnotation")) {
+				this.skipAnnotation = Boolean.valueOf(getProperty("skipAnnotation"));
+			}
+			
+			//First we'll check to see if we have Pipeline-specific download methods
+			//If we don't then all VCF will use the same method.
+			for (Pipeline p : Pipeline.values())
+			{
+				if (hasPropertyAndNotNull(p.toString()+"DownloadMethod")) {
+					this.pipelineDownloadMethods.put(p, DownloadMethod.valueOf(getMandatoryProperty(p.toString()+"DownloadMethod")));
+				}
+				else
+				{
+					// If no download method is defined for a pipeline, we'll set it to null and fill it in later.
+					this.pipelineDownloadMethods.put(p, null);	
+				}
+			}
+
+			
+			//This can be used to fill in gaps where a pipeline-specific value was not given.
+			if (hasPropertyAndNotNull("vcfDownloadMethod")) {
+				this.vcfDownloadMethod = getMandatoryProperty("vcfDownloadMethod");
+				// set any pipeline download method to the defaule of vcfDownloadMethod
+				for (Pipeline p : this.pipelineDownloadMethods.keySet().stream().filter(k -> this.pipelineDownloadMethods.get(k) == null).collect(Collectors.toList()))
+				{
+					this.pipelineDownloadMethods.put(p, DownloadMethod.valueOf(this.vcfDownloadMethod));
+				}
+			}
+			//If the user has set a separate BAM download method, use that. Otherwise, use the same as for VCF.
+			if (hasPropertyAndNotNull("bamDownloadMethod")) {
+				this.bamDownloadMethod = getMandatoryProperty("bamDownloadMethod");
+			}
+			else
+			{
+				this.bamDownloadMethod = this.vcfDownloadMethod;
+			}
+			// By now, we should have everything set up with a downloadMethod. If BAM has not download method, something went wrong and the workflow should terminate.
+			if (!this.skipDownload && (this.bamDownloadMethod == null || this.bamDownloadMethod.trim().equals("")))
+			{
+				throw new RuntimeException("bamDownloadMethod is null/empty! This means that you did not set bamDownloadMethod in the INI file, or you did not set vcfDownloadMethod. At least ONE of these MUST be set.");
 			}
 			
 			this.tumourCount = Integer.valueOf(this.getMandatoryProperty(JSONUtils.TUMOUR_COUNT));
@@ -211,7 +277,7 @@ public abstract class BaseOxoGWrapperWorkflow extends AbstractWorkflowDataModel 
 				this.objectToFilenames.put(tInf.getBamTumourObjectID(), tInf.getTumourBAMFileName());
 				this.objectToFilenames.put(tInf.getBamTumourIndexObjectID(), tInf.getTumourBamIndexFileName());
 				
-				if (this.downloadMethod != null && !this.downloadMethod.equals("") && this.downloadMethod.equals(DownloadMethod.gtdownload.toString()))
+				if (this.vcfDownloadMethod != null && !this.vcfDownloadMethod.equals("") && this.vcfDownloadMethod.equals(DownloadMethod.gtdownload.toString()))
 				{
 					tInf.setTumourBamGNOSRepoURL ( this.getMandatoryProperty(JSONUtils.TUMOUR_BAM_DOWNLOAD_URL+"_"+i) );
 				}
@@ -222,9 +288,11 @@ public abstract class BaseOxoGWrapperWorkflow extends AbstractWorkflowDataModel 
 				{
 					for (VCFType v : VCFType.values())
 					{
-						if (p == Pipeline.muse && v != VCFType.snv)
+						if ( (p == Pipeline.muse && v != VCFType.snv)
+							|| (p == Pipeline.smufin && v != VCFType.indel))
 						{
-							//do nothing when MUSE and non-SNV, because MUSE will ONLY have SNV.
+							// do nothing when MUSE and non-SNV OR when smufin and non-INDEL,
+							// because MUSE will ONLY have SNV, and SMuFin will only have INDEL.
 						}
 						else
 						{
@@ -252,6 +320,9 @@ public abstract class BaseOxoGWrapperWorkflow extends AbstractWorkflowDataModel 
 									break;
 								case muse:
 									vInfo.setPipelineGnosID(this.museGnosID);
+									break;
+								case smufin:
+									//smufin has no GNOS ID so don't need to do anything. 
 									break;
 							}
 							//Only add this object if there is an actual file name present.
@@ -325,26 +396,6 @@ public abstract class BaseOxoGWrapperWorkflow extends AbstractWorkflowDataModel 
 				this.refFile = getProperty("refFile");
 			}
 			
-			if (hasPropertyAndNotNull("skipDownload")) {
-				this.skipDownload = Boolean.valueOf(getProperty("skipDownload"));
-			}
-			
-			if (hasPropertyAndNotNull("skipUpload")) {
-				this.skipUpload = Boolean.valueOf(getProperty("skipUpload"));
-			}
-			
-			if (hasPropertyAndNotNull("skipOxoG")) {
-				this.skipOxoG = Boolean.valueOf(getProperty("skipOxoG"));
-			}
-			
-			if (hasPropertyAndNotNull("skipVariantBam")) {
-				this.skipVariantBam = Boolean.valueOf(getProperty("skipVariantBam"));
-			}
-			
-			if (hasPropertyAndNotNull("skipAnnotation")) {
-				this.skipAnnotation = Boolean.valueOf(getProperty("skipAnnotation"));
-			}
-			
 			if (hasPropertyAndNotNull("gnosMetadataUploadURL")) {
 				this.gnosMetadataUploadURL = getProperty("gnosMetadataUploadURL");
 			}
@@ -354,16 +405,25 @@ public abstract class BaseOxoGWrapperWorkflow extends AbstractWorkflowDataModel 
 			}
 			
 			//System.out.println("DEBUG: downloadMethod: "+this.downloadMethod);
-			if (this.downloadMethod != null && !this.downloadMethod.equals("") && this.downloadMethod.equals(DownloadMethod.gtdownload.toString()))
+			if (this.vcfDownloadMethod != null && !this.vcfDownloadMethod.equals("") && this.vcfDownloadMethod.equals(DownloadMethod.gtdownload.toString()))
 			{
 				System.out.println("DEBUG: Setting gtdownload-specific config values");
 				this.sangerGNOSRepoURL = this.getMandatoryProperty(JSONUtils.SANGER_DOWNLOAD_URL);
 				this.broadGNOSRepoURL = this.getMandatoryProperty(JSONUtils.BROAD_DOWNLOAD_URL);
 				this.dkfzEmblGNOSRepoURL = this.getMandatoryProperty(JSONUtils.DKFZ_EMBL_DOWNLOAD_URL);
 				this.museGNOSRepoURL = this.getMandatoryProperty(JSONUtils.MUSE_DOWNLOAD_URL);
+				this.gtDownloadVcfKey = this.getMandatoryProperty("gtDownloadVcfKey");
+			}
+			else if ((this.vcfDownloadMethod != null && !this.vcfDownloadMethod.equals("") && this.vcfDownloadMethod.equals(DownloadMethod.filesystemCopy.toString()))
+					||(this.bamDownloadMethod != null && !this.bamDownloadMethod.equals("") && this.bamDownloadMethod.equals(DownloadMethod.filesystemCopy.toString())))
+			{
+				this.fileSystemSourcePath = this.getMandatoryProperty("fileSystemSourcePath");
+			}
+			
+			if (this.bamDownloadMethod != null && !this.bamDownloadMethod.equals("") && this.bamDownloadMethod.equals(DownloadMethod.gtdownload.toString()))
+			{
 				this.normalBamGNOSRepoURL = this.getMandatoryProperty(JSONUtils.NORMAL_BAM_DOWNLOAD_URL);
 				this.gtDownloadBamKey = this.getMandatoryProperty("gtDownloadBamKey");
-				this.gtDownloadVcfKey = this.getMandatoryProperty("gtDownloadVcfKey");
 			}
 			
 		} catch (Exception e) {