+++ src/main/java/com/github/seqware/CommonPredicates.java
@@ -11,6 +11,7 @@ public interface CommonPredicates {
 	static final Predicate<VcfInfo> isBroad = p -> p.getOriginatingPipeline() == Pipeline.broad;
 	static final Predicate<VcfInfo> isDkfzEmbl = p -> p.getOriginatingPipeline() == Pipeline.dkfz_embl;
 	static final Predicate<VcfInfo> isMuse = p -> p.getOriginatingPipeline() == Pipeline.muse;
+	static final Predicate<VcfInfo> isSmufin = p -> p.getOriginatingPipeline() == Pipeline.smufin;
 	
 	static final Predicate<VcfInfo> isIndel = p -> p.getVcfType() == VCFType.indel;
 	static final Predicate<VcfInfo> isSnv = p -> p.getVcfType() == VCFType.snv;